requirements.txt
hydra-core==1.3.1
numpy==1.23.5
scikit-learn==1.2.0
sentencepiece==0.1.99
seqeval==1.2.2
torch
tqdm==4.64.0
transformers==4.35.2